requirements.txt
asn1crypto==0.24.0
certifi==2018.10.15
cffi==1.7.0
chardet==3.0.4
Click==7.0
cryptography==2.4.2
flask>=0.12.3
future==0.17.0
idna==2.5
itsdangerous==1.1.0
Jinja2==2.10
line-bot-sdk==1.8.0
MarkupSafe==1.1.0
pycparser==2.19
pyOpenSSL==18.0.0
requests==2.20.0
six==1.11.0
urllib3==1.22
Werkzeug==0.14.1